To decimal

+++ Joanna_Laurence/src/main/java/org/uva/sea/ql/parser/elements/types/Money.java
@@ -4,22 +4,23 @@ import org.uva.sea.ql.parser.elements.ASTNode;
 import org.uva.sea.ql.parser.elements.TraverseType;
 import org.uva.sea.ql.traverse.Traverse;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public class Money extends ASTNode {
     private String currency;
-    private BigInteger amount;
+    private BigDecimal amount;
 
     public Money(String currency, String amount) {
         this.currency = currency;
-        this.amount = new BigInteger(amount);
+        this.amount = new BigDecimal(amount);
     }
 
     public String getCurrency() {
         return currency;
     }
 
-    public BigInteger getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
